+++ lib/core/l10n/app_bo.arb
@@ -717,6 +717,7 @@
   "group_links_title": "འབྲེལ་ཐག",
   "group_and_more_links": "དང་འབྲེལ་ཐག་གཞན་ {count}",
   "group_practice_with_us": "ང་ཚོ་དང་མཉམ་དུ་སྒོམ་གྱི།",
+  "series_practicing_with_group": "{groupName} དང་མཉམ་དུ་ཉམས་ལེན་བྱེད",
   "group_join_to_contribute": "བསགས་གྲངས་སྐྱེལ་ཆེད་ཞུགས།",
   "group_accumulator_join_error": "བསགས་གྲངས་ལ་ཞུགས་མ་ཐུབ། ཡང་བསྐྱར་ཚོད་ལྟ་གནང་རོགས།",
   "group_accumulator_participants": "མཉམ་ཞུགས་པ་ {count}",
