+++ lib/core/l10n/app_en.arb
@@ -794,6 +794,14 @@
   "group_links_title": "Links",
   "group_and_more_links": "and {count} more links",
   "group_practice_with_us": "Practice with us",
+  "series_practicing_with_group": "Practicing with {groupName}",
+  "@series_practicing_with_group": {
+    "placeholders": {
+      "groupName": {
+        "type": "String"
+      }
+    }
+  },
   "group_change_practice_title": "Change group practice",
   "group_change_practice_message": "You are already practicing this plan with another group. Would you like to change your practice group?",
   "group_join_to_contribute": "Join to contribute",
