+++ lib/core/l10n/app_hi.arb
@@ -783,6 +783,7 @@
   "group_links_title": "लिंक",
   "group_and_more_links": "और {count} और लिंक",
   "group_practice_with_us": "हमारे साथ अभ्यास करें",
+  "series_practicing_with_group": "{groupName} के साथ अभ्यास",
   "group_join_to_contribute": "योगदान के लिए शामिल हों",
   "group_accumulator_join_error": "संचय में शामिल होने में असमर्थ। कृपया पुनः प्रयास करें।",
   "group_accumulator_participants": "{count} प्रतिभागी",
