+++ lib/core/l10n/app_mn.arb
@@ -783,6 +783,7 @@
   "group_links_title": "Холбоосууд",
   "group_and_more_links": "болон дахиад {count} холбоос",
   "group_practice_with_us": "Бидэнтэй хамт дадлага хий",
+  "series_practicing_with_group": "{groupName}-тай хамт дадлага хийж байна",
   "group_join_to_contribute": "Хувь нэмэр оруулахын тулд нэгдэх",
   "group_accumulator_join_error": "Хуримтлалд нэгдэх боломжгүй байна. Дахин оролдоно уу.",
   "group_accumulator_participants": "{count} оролцогч",
