+++ lib/features/home/data/datasource/home_local_datasource.dart
@@ -47,7 +47,8 @@ class HomeLocalDatasource {
   String featuredSeriesKey(String language, int limit) =>
       'featured_series:v3:$language:$limit';
   String seriesListKey(String language) => 'series_list:$language';
-  String seriesByIdKey(String language, String id) => 'series:$language:$id';
+  String seriesByIdKey(String language, String id) =>
+      'series:v2:$language:$id';
   String todayEventsKey(String language) => 'today_events:$language';
   String verseOfDayKey(String language) => 'verse_of_day:$language';
   String routineInfoKey(String userId) => 'routine_info:$userId';
