+++ lib/features/practice/presentation/widgets/practice_plan_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_pecha/core/constants/app_assets.dart';
 import 'package:flutter_pecha/core/theme/app_colors.dart';
 import 'package:flutter_pecha/core/widgets/responsive_cover_image.dart';
 import 'package:flutter_pecha/features/home/domain/entities/series.dart';
+import 'package:flutter_pecha/features/home/presentation/widgets/series_plan_card_widgets.dart';
 import 'package:flutter_pecha/features/plans/data/utils/plan_date_format.dart';
 
 class PracticePlanCard extends StatelessWidget {
@@ -16,6 +16,9 @@ class PracticePlanCard extends StatelessWidget {
   final Series series;
   final VoidCallback onTap;
   final double width;
+
+  static const _metaFontSize = 12.0;
+  static const _titleFontSize = 14.0;
 
   @override
   Widget build(BuildContext context) {
@@ -61,7 +64,7 @@ class PracticePlanCard extends StatelessWidget {
               child: Text(
                 series.title,
                 style: const TextStyle(
-                  fontSize: 14,
+                  fontSize: _titleFontSize,
                   fontWeight: FontWeight.w600,
                   height: 1.3,
                 ),
@@ -69,15 +72,36 @@ class PracticePlanCard extends StatelessWidget {
                 overflow: TextOverflow.ellipsis,
               ),
             ),
-            if (dateRange != null || series.enrolledCount > 0)
+            if (series.partner != null)
               Padding(
                 padding: const EdgeInsets.fromLTRB(8, 0, 8, 8),
-                child: _PracticePlanMetaRow(
-                  dateRange: dateRange,
-                  enrolledCount: series.enrolledCount,
-                  fontSize: 12,
-                  secondaryColor:
-                      Theme.of(context).colorScheme.onSurfaceVariant,
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Expanded(
+                      child: SeriesPlanPartnerRow(
+                        partner: series.partner!,
+                        fontSize: _metaFontSize,
+                        avatarSize: 28,
+                      ),
+                    ),
+                    if (series.enrolledCount > 0) ...[
+                      const SizedBox(width: 8),
+                      SeriesPlanEnrolledCount(
+                        count: series.enrolledCount,
+                        fontSize: _metaFontSize,
+                      ),
+                    ],
+                  ],
+                ),
+              )
+            else if (dateRange != null || series.enrolledCount > 0)
+              Padding(
+                padding: const EdgeInsets.fromLTRB(8, 0, 8, 8),
+                child: SeriesPlanMetaRow(
+                  series: series,
+                  fontSize: _metaFontSize,
                 ),
               ),
           ],
@@ -88,43 +112,5 @@ class PracticePlanCard extends StatelessWidget {
 
   static String? _formatSeriesDateRange(Series series) {
     return PlanDateFormat.formatRangeOrNull(series.startDate, series.endDate);
-  }
-}
-
-class _PracticePlanMetaRow extends StatelessWidget {
-  const _PracticePlanMetaRow({
-    required this.dateRange,
-    required this.enrolledCount,
-    required this.fontSize,
-    required this.secondaryColor,
-  });
-
-  final String? dateRange;
-  final int enrolledCount;
-  final double fontSize;
-  final Color secondaryColor;
-
-  @override
-  Widget build(BuildContext context) {
-    final textStyle = TextStyle(fontSize: fontSize, color: secondaryColor);
-
-    return Row(
-      children: [
-        if (dateRange != null)
-          Expanded(
-            child: Text(
-              dateRange!,
-              style: textStyle,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-          ),
-        if (enrolledCount > 0) ...[
-          Icon(AppAssets.usercard, size: fontSize + 2, color: secondaryColor),
-          const SizedBox(width: 4),
-          Text('$enrolledCount', style: textStyle),
-        ],
-      ],
-    );
   }
 }
